task 100-prime_factor.c commit

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-	int num = 612852475143;
+	long int num = 612852475143;
 	int pf = 2;
 	int largest = 0;
